requirements/production.txt
# PRECAUTION: avoid production dependencies that aren't in development

-r ./base.txt

gunicorn==19.8.1  # https://github.com/benoitc/gunicorn
psycopg2==2.7.4 --no-binary psycopg2  # https://github.com/psycopg/psycopg2
Collectfast==0.6.2  # https://github.com/antonagestam/collectfast

# Django
# ------------------------------------------------------------------------------
django-storages[boto3]==1.6.6  # https://github.com/jschneier/django-storages
django-anymail[mailgun]==3.0  # https://github.com/anymail/django-anymail
django_content_type
